app/requirements.txt
fastapi==0.73.0
uvicorn==0.17.4
pydantic[email]
motor==2.5.1
cryptography
requests
PyJWT
pyotp
